+++ example/dio_sse_example.dart
@@ -3,11 +3,11 @@ import 'package:dio_sse/dio_sse.dart';
 
 Future<void> main() async {
   final dio = Dio();
-  final sse = EventSource.from(dio, url: "https://sse.dev/test");
-  sse.start();
+  final sse = EventSource.from(dio, url: 'https://sse.dev/test');
   sse.onAnyMessage().listen((event) {
     print(event);
   });
-  await Future.delayed(Duration(milliseconds: 10000));
+  await sse.start();
+  await Future<void>.delayed(const Duration(seconds: 10));
   await sse.dispose();
 }